requirements.txt
fastapi
uvicorn
jinja2
httpx
pdfplumber
python-docx
requests
transformers
torch
python-multipart
dotenv